+++ packages/agent/src/execution-mode.ts
@@ -1,4 +1,4 @@
-import { type Adapter, CODEX_MODE_PRESETS } from "@posthog/shared";
+import { CODEX_MODE_PRESETS, type ExecutionMode } from "@posthog/shared";
 import { ALLOW_BYPASS } from "./utils/common";
 
 export interface ModeInfo {
@@ -68,7 +68,7 @@ export const CODEX_NATIVE_MODES = ["auto", "read-only", "full-access"] as const;
 export type CodexNativeMode = (typeof CODEX_NATIVE_MODES)[number];
 
 /** Union of all permission mode IDs across adapters */
-export type PermissionMode = CodeExecutionMode | CodexNativeMode;
+export type PermissionMode = ExecutionMode;
 
 export function isCodexNativeMode(mode: string): mode is CodexNativeMode {
   return (CODEX_NATIVE_MODES as readonly string[]).includes(mode);
@@ -80,36 +80,4 @@ export function getAvailableCodexModes(): ModeInfo[] {
   return ALLOW_BYPASS
     ? [...CODEX_MODE_PRESETS]
     : CODEX_MODE_PRESETS.filter((m) => m.id !== "full-access");
-}
-
-// The cloud task-run API validates initial_permission_mode against a per-adapter
-// enum that is narrower than the local pickers: codex accepts only its native
-// modes (notably not "plan", which the local app-server adapter synthesizes).
-// Degrade to the nearest mode with the same permission ceiling.
-const CODEX_CLOUD_MODE_FALLBACKS: Record<
-  Exclude<CodeExecutionMode, "auto">,
-  CodexNativeMode
-> = {
-  default: "auto",
-  acceptEdits: "auto",
-  plan: "read-only",
-  bypassPermissions: "full-access",
-};
-
-const CLAUDE_CLOUD_MODE_FALLBACKS: Record<
-  Exclude<CodexNativeMode, "auto">,
-  CodeExecutionMode
-> = {
-  "read-only": "plan",
-  "full-access": "bypassPermissions",
-};
-
-export function resolveCloudInitialPermissionMode(
-  adapter: Adapter,
-  mode: PermissionMode,
-): PermissionMode {
-  if (adapter === "codex") {
-    return isCodexNativeMode(mode) ? mode : CODEX_CLOUD_MODE_FALLBACKS[mode];
-  }
-  return isCodeExecutionMode(mode) ? mode : CLAUDE_CLOUD_MODE_FALLBACKS[mode];
 }
